Corporate_Email/requirement.txt
Django==1.10.1
pydns==2.3.6
validate-email==1.3
